+++ src/applications/content-drop-folders-app/content-drop-folders-app.module.ts
@@ -22,7 +22,7 @@ import { KalturaPrimeNgUIModule } from '@kaltura-ng/kaltura-primeng-ui';
 import { AutoCompleteModule } from '@kaltura-ng/kaltura-primeng-ui/auto-complete';
 import { TagsModule } from '@kaltura-ng/kaltura-ui/tags';
 import { PopupWidgetModule } from '@kaltura-ng/kaltura-ui/popup-widget';
-import { ContentSharedModule } from 'app-shared/content-shared/content-shared.module';
+import { EntriesModule } from 'app-shared/content-shared/entries/entries.module';
 
 import { ContentDropFoldersComponent } from './content-drop-folders.component';
 import { DropFoldersComponentsList } from './drop-folders-components-list';
@@ -54,7 +54,7 @@ import { FiltersModule } from '@kaltura-ng/mc-shared/filters/filters.module';
     FormsModule,
     ReactiveFormsModule,
     StickyModule,
-    ContentSharedModule,
+      EntriesModule,
     RouterModule.forChild(routing),
     FiltersModule
   ],